requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn==1.6.1
joblib>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
tensorflow>=2.14.0
Pillow>=10.0.0
matplotlib>=3.8.0
shap>=0.45.0
pytest>=8.0.0
